Added support for Mozilla DeepSpeech
Added support for Mozilla DeepSpeech. This will require some updates to
populate.py when we fix it. A profile.yml file set to use deepspeech
as the active listener will have a section like this:

active_stt:
  engine: deepspeech-stt
deepspeech:
  model: '/home/user/models/output_graph.pb'
  alphabet: '/home/user/models/alphabet.txt'
  language_model: '/home/user/models/lm.binary'
  trie: '/home/user/models/trie'

or similar, depending on the location where the the models were built.

Deepspeech will build on the Raspberry Pi, but was not truly usable on
it the last time I checked.