Added ability to choose audio engine
Added the ability to choose the audio engine from the current
two choices of alsa or pyaudio. Then list available input
and output devices and test to make sure they are working.

Continued to improve the colorization by defining functions
for specific purposes.

Because I am now accessing some of the Naomi packages directly
to query options and test settings, I had to change the way
this program is launched. It can still be invoked just before
running Naomi by launching it as:
[~/Naomi]$ ./Naomi.py --repopulate
it will also run automatically if Naomi can't find a profile.
If you want to run it alone, you should run it as
[~/Naomi]$ ./Populate.py
rather than as
[~/Naomi]$ naomi/populate.py