Automate maintenance of translation files
This commit adds a new program called "update_translations.py"
which is used to maintain Naomi's .po translation files. This
should be run whenever you add or change a line of quoted text
that will be used as output for Naomi, or if someone else has
committed a change and you are maintaining a specific language
file. This will create a space in the .po file where a translator
can manually enter the correct translation in the language.

usage: update_translations.py [-h] [-l LANGUAGE [LANGUAGE ...]] [-d]

Update .pot files to receive updated translation strings

optional arguments:
  -h, --help            show this help message and exit
  -l LANGUAGE [LANGUAGE ...], --language LANGUAGE [LANGUAGE ...]
  -d, --debug           Show debugging info

When the program is run, it attempts to update the .po files in
naomi/data/locale, then it runs through all the plugins and
updates the language files for any plugin that contains a "locale"
directory.

If you request a specific language or set of languages, it will
only update or create that language or those languages. If you
don't list any specific languages, it will work from the list of
language files in naomi/data/locale.