Fixed several issues with populate.py
Added options for Wit.AI so if that STT engine is selected, it
will ask for a token and not default to the pocketsphinx questions.

Also fixed an issue where it was writing the language choice as
language: !!python/unicode 'en-US'
which was causing problems for the yaml library.

Also fixed an issue where it was asking for an email password
even if you did not enter an email address.