This commit adds a new program called "update_translations.py" which is used to maintain Naomi's .po translation files, which are the files that the compile_translations.py program uses to build the .mo files that are actually used.
This should be run whenever a developer adds or changes a line
of quoted text which will be output to the user, or if you are
maintaining a translation, you can use it to make sure that
all quotes in the source code have been pulled into the .po file.
After running this program, the .po files will be updated with
new lines where new translations can be entered.

usage: update_translations.py [-h] [-l LANGUAGE [LANGUAGE ...]] [-d]

optional arguments:
-h, --help - show a help message and exit
-l, --language LANGUAGE[, LANGUAGE...]] - create translations for
language (ie. en-US)
-d, --debug - Show debugging info

When the program is run, it attempts to update or create the .po
files in naomi/data/locale, then it walks through the plugins
directory and updates or creates the language files for any
plugin that contains a "locale" directory.

If you pass in a specific language or set of languages, it will
only update or create that language or those languages. If you
don't list any specific languages, it will work from the list
of language files in naomi/data/locale.