Fix issue where dictionary is lower case
When the dictionary is lower case, but the list of words to be
compiled is in upper case, each letter is individually flagged
as "Input symbol not found". At that point, the error handler
in sphinxvocab.py is supposed to kick in and attempt to convert
the list of words to lower case and feed it back in. If the
error occurs at that point, the program dies with an "Input
symbol not found" error and a stack trace.

At some point, while updating the new pocketsphinx module, I
had run into this problem on the second run-through and decided
it would be better to just skip whatever line was having the
issue, since it was not related to the functioning of Naomi.

I changed the error into a continue. This meant that if I used
a lower-case dictionary, none of the words would be compiled.

When I started using the english dictionary that is currently
bundled with the git version of pocketsphinx, it was in lower
case, which no longer worked with my code. I dealt with this
by checking to see which version of pocketsphinx is installed
and upper or lower-casing the dictionary accordingly. Of
course, the version of pocketsphinx has exactly nothing to do
with whether the chosen dictionary is upper or lower case.

A dictionary should be either all upper case or all lower case,
since otherwise Phonetisaurus might end up associating the upper
and lower case letters with different sounds.

This issue was caught by gospodima, who was attempting to use
the older version of phonetisaurus from the original jasper
documentation with a lower-case german dictionary.

This commit reverses those specific changes to sphinxvocab.py
and g2p.py and allows words to be compiled against phoneme
dictionaries regardless of whether the dictionaries are upper
or lower case.