Add gettext to naomi-setup.sh
This is a few additional tweaks to naomi-setup.sh. First, I
updated the header to reflect that python would be installed
to the ~/.naomi/local/bin directory and not to the ~/local/bin
directory. That was a change I made just to keep all the
naomi support files together in one location and make it clear
what they are for.

Then I changed the instructions about needing to install
libssl1.0-dev to libssl-dev just to match the actual library
that seems to work.

I added gettext to the list of apt modules being installed so
that the system is ready to run "compile_translations.sh" later,
and added some text explaining to the user what is being done
in case they receive a sudo password prompt.

Added a line telling the user to use the "./Naomi" command instead
of "./Naomi.py" to start Naomi. Added --repopulate to the ./Naomi
command to force repopulation of the profile even if one already
exists.