Adjusting setup script & python requirements (#122)
* Adjusting setup script & python requirements

* Update README.md

* Fixing Setup & Requirments

* Resolving Readme Conflict

* Updated pull request template

* Update bug report template

* Updating feature request template

* Updating readme

* Fixes to naomi-setup.py

The naomi-setup.py script installs its own version of python
locally under user permissions.

The version of pip was old and attempting to download package
files without using ssl, which appears to no longer be supported
by PyPi.org, so uusing it to process python-requirements.txt was
just failing. I updated that, then I was getting error messages
about failing to build the _ssl module when building python, so
was still unable to download packages from pypi.org. Moved the
apt installs to before installing python and installed the
libssl1.0-dev. Upgraded python from 2.7.3 to 2.7.13.

Saved the directory that the naomi-setup.sh file lives in since
that should be the location of the Naomi install. Return to
that directory after installing python, pip, etc. and don't just
assume the directory is called ~/Naomi.

After returning, you need to run Naomi using the copy of python
you just installed. You have three choices there. Ensure that
/usr/bin/env always finds your python2 first, change the crunch
bang line in Naomi.py, or create a shell script that calls
Naomi.py with the correct version of python. I went for option
three.

* Add gettext to naomi-setup.sh

This is a few additional tweaks to naomi-setup.sh. First, I
updated the header to reflect that python would be installed
to the ~/.naomi/local/bin directory and not to the ~/local/bin
directory. That was a change I made just to keep all the
naomi support files together in one location and make it clear
what they are for.

Then I changed the instructions about needing to install
libssl1.0-dev to libssl-dev just to match the actual library
that seems to work.

I added gettext to the list of apt modules being installed so
that the system is ready to run "compile_translations.sh" later,
and added some text explaining to the user what is being done
in case they receive a sudo password prompt.

Added a line telling the user to use the "./Naomi" command instead
of "./Naomi.py" to start Naomi. Added --repopulate to the ./Naomi
command to force repopulation of the profile even if one already
exists.