Starting Python3 conversion
Rewrite naomi-setup.sh to install and link to python3. Also
fixed an issue at the end of the setup where it writes the final
./Naomi file where it was not putting a literal $@ at the end
of the line, so not properly passing command line arguments through.

Added a line to .gitignore to ignore the ./Naomi file created by
naomi-setup.sh

Set crunch-bang line in Naomi.py to use python3 instead of python2.

Changed ./naomi/__main__.py to set *_MIC globals from application
rather than just importing them, because the relative import did
not work. There are a few ways to fix that, but this seemed easiest
for right now.

./naomi/i18n.py, ./naomi/mic.py, ./plugins/tts/flite-tts/flite.py,
./plugins/stt/pocketsphinx-stt/g2p.py,
./plugins/stt/pocketsphinx-stt/sphinxplugin.py,
./plugins/stt/pocketsphinx-stt/sphinxvocab.py
Changed some encoding, fixing new errors around str to byte
conversions.

./plugins/stt/pocketsphinx-stt/sphinxplugin.py
Changed the import from

import pocketsphinx

to

from pocketsphinx import pocketsphinx

also added "from sphinxbase.sphinxbase import *" because otherwise
the line "config = pocketsphinx.Decoder.default_config()"
did not work. I'm not sure why that is necessary. I am currently
running the git version of pocketsphinx, so I don't know if there
are different problems around earlier versions, or why you have
to import pocketsphinx.pocketsphinx now instead of just
pocketsphinx.

./plugins/speechhandler/joke/data/en-US.txt
restored file from Jasper original file. I'm not sure why this was
removed. I did not restore the german joke file. I don't think
there has ever been a french joke file.

./plugins/speechhandler/weather/weather.py
discovered that weather is reading the location from profile.yml at
weather:
  location:
also converting the location to woeid, which can be stored at
weather:
  woeid:
so I added an info line stating what the woeid is so it can
be put in the profile.yml file and the conversion step can be
skipped.

./populate.py
Changed the location of the weather plugin "location" variable
from

location:

to

weather:
  location:

My copy of Naomi is now running on python3 using pocketsphinx
and flite. Overall, this was a lot easier than I thought it was
going to be, and mostly just involved running Naomi.py and
waiting for it to crash with an error about needing a byte
object and finding a string object (add b before the quotes
or add ".encode('utf-8')") or needing a string and finding a
byte object (add .decode('utf-8') to the end).