Fixed a few issues with Naomi 2.2
Added the generated Naomi file that uses the local python to the
.gitignore list

Modified naomi-setup.sh and added libncurses5-dev to the list
of apt packages to install

Removed nano and git from python-requirements.txt (duplicating
PR # 131 by redragonx)

Added a check for audio engines to populate.py that runs directly
after the language selection, so the user doesn't get to the end
and suddenly discover they can't complete the install.

Fixed an issue with sphinxvocab.py where temporary files were
not being removed. This was for debugging purposes.