Add Audio Transcript
Added the option to view the transcript of what naomi is hearing
and saying. This is mostly for debugging purposes.

Added the command-line option "--print_transcript"
Added the profile.yml option "print_transcript:"

The command-line option overrides the profile option.

Prints things naomi says with two less-than's: "<<"
Prints things naomi hears in passive mode with a single
greater-than: ">"
Prints things naomi hears in active mode with a double
greater-than: ">>"

This could be a privacy issue if someone's conversation is
transcribed when they aren't expecting it to be, which is
why I made it opt-in. We might want to print or say a warning
that the operator is responsible for warning people that what
they say near Naomi might be recorded into a text log.