Additional options for naomi-setup.sh
This gives the user three possible ways to install Naomi.
Unfortunately, all have their own benefits and issues,
mostly due to the fact that Python itself is still in
transition from version 2 to version 3.

PROS
This is my current recommendation since the option of downloading
and compiling a special version of python just for Naomi turned
out to be problematic, and because this way the python
interpreter can be kept up to date through the use of the
distro's package manager.

Once the user has activated the virtual environment, they can
just use 'pip' and 'python' to manage the environment.

CONS
I have had the experience of virtualenvs breaking when the distro
moves from one minor version of python to another, but then I am
on a rolling distro, and that is less likely to happen with
more stable distros. It requires installing python and python-pip
in addition to python3 and python3-pip (pip3) because if you use
virtualenvwrapper with python3 right now, you have to specify the
python3 interpreter with the VIRTUALENVWRAPPER_PYTHON and
VIRTUALENVWRAPPER_VIRTUALENV environment variables.

If the user is using Pocketsphinx and did not have python3
installed while running through the install procedure, then they
may need to "workon Naomi" and re-install the pocketsphinx,
sphinxbase and cmuclmtk python modules, which adds complexity.

PROS
This leaves the user's copy of python untouched. It creates a
stable, independent environment for Naomi to work from so it
is less likely that a distro update will break stuff.

CONS
First, we had problems right away with Ubuntu 18.10 and
derivative distros. The gcc compiler included with them does
not appear to be compatible with building python3, so these
users really need to use the pre-built python3 binaries.

Building python3 was also complex in that there are a lot of
additional dev library requirements when building packages
like "ssl" and "curses" that are required for Naomi.

This copy of python will not receive updates when the user
updates the system using the standard package manager, meaning
that exploits are less likely to get fixed in a timely fashion.

This is fairly complex, and I have implemented redragonx's
signature and checksum checks. Any minor change to any of the
setup files would require updating this script since the
signature and checksum hashes are embedded.

When installing or updating packages, the user and any scripts
have to know to use the ~/.naomi/local/bin/pip command rather
than the system "pip" command.

PROS
Easy to setup, easy to use, user and scripts don't have to
remember to do anything when updating and installing libraries
and can just use 'pip3' and 'python3' commands.

CONS
More and more software is written in python, so there are a
lot of system dependancies you won't necessarily be aware of.
Since we are installing specific versions of pip modules
and not using they system package manager, there is a very
real possibility that this might break a dependancy for
another program on the user's system.