Add post-install instructions
There are still things that users need to know and be aware of
after running this installer. If they choose the first option,
then they will need to add the virtualenv initialization script
to their shell startup script. If they choose the second option,
they need to be aware that they need to use Naomi's special
copy of python when installing modules or running tests.