Fixed several issues with populate
We had a few issues with populate. First, it was crashing if
the profile.yml file did not exist because I was using the profile
module to load the profile and it wasn't checking for the file
first.

Second, we needed to get rid of the question asking what your
location was for the weather plugin. Both the weather service
and the geocoding service are now defunct, so if you actually
tried to answer the question you would be unable to continue.

Added in some code that is supposed to check devices for
compatibility. It doesn't seem to work very well, but it's
there for others to look at.

Added some code for testing and accepting the voice selection
when using flite.

Reordered some of the questions so the device configuration stuff
comes first and questions that are less likely to cause issues
come after.