Exit should turn off music (#180)
This looks good to me. I have never used the mpd control before and also don't have a spotify account, so from the documentation I can't really test this. I've actually never spoken with anyone who has used the mpd plugin before, so it's exciting that it appears to still work. There is an issue with the python-mpd library in python3, however. It is detailed in issue #167 

It looks like a minor change that should do just what you say