Fix unkown item (#178)
This looks good. I have no real way of testing this, but it does no appear to break anything either through unit testing or when running Naomi. I don't exactly understand how it works, look like it is setting the default for title, artist and album to "default" which I assume works as a wild card with Spotify?