Passive Listen
This allows the user to set an "active_listen" value in profile.yml
to True (or "Yes") in order to change Naomi's listening behavior.

Normally, you speak Naomi's keyphrase, Naomi responds with a high
beep noise, then you speak your query, Naomi response with a low
beep noise, then processes the request.

Using passive listen, Naomi listens for your voice, then scans the
entire block of audio for the keyword. If it detects the keyword,
it passes the audio to the active listener, then finally returns
the transcription.

This pull request also fixes the issue that at some point the low
beep got replaced with the high beep. I think it was when I was
working on the Voice Activity Detector plugin.