Adding wwis_weather plugin
Here I am adding the new wwis_weather plugin. This will be used
to develop the new form description structures that I hope to be
able to give to plugin developers as a way of describing settings
they need to get from the user. In this case, I need a "cityId"
variable from the WWIS database for where the weather should be
retrieved if the user makes a request without mentioning a location
ie: "Is it going to rain tomorrow?" or "Tell me the weather
forecast".

Getting this cityId field requires drilling down into the WWIS
location database. First you choose a country, then that country
can either contain cities, regions that countain cities, a
combination of both regions and cities, or no city information
(in which case no weather information is available). If you
choose a region in the country, then you have to choose a city.

This is all pretty complex, so it seems like a good test for
robustness.