Fixing google stt plugin (#184)
I have reviewed these changes and they work for me on my Raspberry Pi 3B+ running Raspbian Stretch.

The instructions at https://cloud.google.com/speech-to-text/docs/quickstart-protocol (as noted in the source code) worked for me. The documentation as it currently stands could be updated/improved but it's not strictly necessary, since the current link does eventually get the user to the instructions for creating an account, setting up billing, downloading the json authorization file, and setting the environment variable.

It's nice to have the google stt option available again. Thanks for the help!