A few additional changes to google stt
I noticed that the test requires the environment variable to
exist in order to not be skipped, but then in the main google.py
module, you are getting the location of the file from profile.py.

I think this modification will favor the environment variable over
the profile.py setting (if both are set, the environment variable
wins) in both the google.py and tests/test_googlesttplugin.py.

I haven't had a chance to thoroughly test this yet, but I did
run unittest discovery against it with a valid authorization file
in profile.yml and that worked.