Fix passive listen
When testing today I discovered that using the google stt plugin
no audio was being transcribed due to the fact that that plugin
does not seek a location in the passed file handle, which the
pocketsphinx and deepspeech plugins I tested against before do.
Since it is not a requirement that stt plugins set the file
location, and it is reasonable to assume that the file handle is
currently pointed to the beginning of the file, it seemed
reasonable to fix this in the naomi/mic.py module.

This patch just sets the file pointer back to the beginning of
the file before calling active_stt_engine.transcribe() on the
active file handle when using passive_listen mode.