Unit tests for google stt plugin, updated docs and credentials json (#185)
Google STT can now use a setting in profile.yml instead of an environment variable to store the location of the credentials file. Additionally, the python-requirements.txt file has been updated to automatically pull in the required libraries for this plugin.

In the process of resolving the differences with google.py, also fixed the codacy complaint.