Fixed a codacy issue
Codacy didn't like that I was using a class method for my fake
special_mode method for local_mic.py that was required to be able
to use speechhandlers that make use of special modes while using
the --local option.

While I was in there, I also updated python_requirements.txt to
require the python3 compatible python-mpd2 library instead of the
python-mpd library which is not compatible with python3.