Fix to NaomiSTTTrainer.py
Fixed two problems with the NaomiSTTTrainer.py server. First,
in going through the flake8 testing, I reformatted the code for
generating the audio control into a join statement, but forgot
to put commas at the ends of lines, so the <audio controls was
becoming <audiocontrols which html5 was ignoring.

I also changed the function of the "submit" so that after
submitting a change, you are taken to the next unverified
record. So say the database has 100 recordings in it, and
during the last session you skipped record 50 and then
continued to verify records until 80, then records 81-100
were created the next time you ran Naomi. You will automatically
be taken to record 50 first, then as soon as you verify it,
you will be taken to record 81. If you use the "next" button
to step through all records, though, you will be taken to
record 51.