Removing STT Trainer while I work on it
The STT Trainer is getting flagged by Codacy due to the fact that
it uses a lot of subprocess calls in the training, and the return
from these calls is different from what is being returned by
naomi/run_command.py. I'd like to re-write naomi/run_command.py
which is definitely going to get that pull request flagged by
Codacy, so I'm going to pull that into another small pull request.