Fixed issue with creating encryption key
If you didn't have an encryption key, the process of checking the
profile for a password would fail, which was causing new values
added for the module not to be written to the profile. Naomi would
still start up, but it was impossible to enter and save your
email password for the check_email plugin.