Updated run_command
Updated run_command so that you can specify which output streams
you would like to capture. This is currently only used for getting
the list of voices available to Flite in populate.py, but I am
using it a lot in my STT training stuff.

Another small change is changing the default TTS engine to Flite.