NaomiSTTTrainer
This commit adds the NaomiSTTTrainer.py program. This runs as a
webserver under WSGI. You don't need a webserver to run it, just
run ./NaomiSTTTrainer.py and it should create a webserver
listening on port 8080 on your local machine. This is not intended
to run while you are running Naomi, it is only used for managing
your local database of audio clips and training different
speech to text engines. You should stop the webserver when you
are finished by pressing CTRL-C in the terminal window where it
is running.

If you start it up from within a GUI, it will also attempt to
launch your default browser.

This only works for users who have added the following line to
the Naomi profile:

audiolog:
  save_audio: true

Users will also have to install SphinxTrain:
  git clone https://github.com/cmusphinx/sphinxtrain.git

The interface consists of a page with two tabs
  * Verify transcriptions - this is the page that currently comes
  up by default. Towards the top of this window is an HTML5
  audio control. Play it to hear the audio clip Naomi recorded.
  Below that is a line that reads:
    Computer heard "{what the computer heard}"
  and below that is a series of radio buttons:

    The transcript is correct. I heard the same thing
    The transcript is not correct. This is what I heard:
    This was just noise with no voices
    This was too unclear to understand

  After listening to the clip, choose the appropriate radio
  button. If you heard something different from what Naomi
  heard, then type in your transcription. Naomi should
  automatically remove any punctuation and convert everything
  to upper case when training, so you don't have to follow
  the same convention of writing every word in upper case and
  writing everything in upper case that Naomi does.

  Audio that is just noise but no voices is important to
  capture for training the noise dictionary. Right now
  anything that is too unclear for a human to understand
  will not be used for training.

  * Train STT Engines - This tab has only one button on it
  currently, "Adapt Pocketsphinx". I have designed this
  program to work with Naomi's plugin system, so it should
  be easy to add additional STT trainers, to finally allow
  users to build and train their own models for Julius,
  Kaldi and DeepSpeech after running Naomi for a while using
  either standard PocketSphinx models or a cloud STT system
  like Google Cloud STT.

  If anything goes wrong while adapting your model, error
  messages should appear in the terminal window.

To accomplish this, I have changed the following files:

naomi/application.py - moved the check for populate.py back
into naomi/application.py because I was sometimes getting infinite
loops with it in naomi/commandline.py. To move the call to populate
back into application when the profile.yml file does not exist,
I am having profile.py create a flag that lets us know that the
profile was missing when we bubble back up to application.py.

naomi/brain.py - Made use of a new standard phrases file created
by Pocketsphinx_adapt. As the user trains the speech recognition,
the actual words used but not part of the "get_phrases" system
are put into this file to improve word recognition rates.

naomi/commandline.py - converted the commandline program into an
object. This was due to having trouble with it not initializing
correctly. This all has to do with the language setting. You want
the language to be set correctly before you can parse the i18n
translations, so you need to make sure the language is set up. At
the same time, if the language is currently set, you want to respect
the current settings when asking which language to use. The main
difference in converting this library into an object is just the
way it is called. Instead of doing:

  from naomi import commandline as interface
  interface.do_something()

You now use:

  from naomi import commandline
  interface = commandline.commandline()
  interface.do_something()

naomi/i18n.py - added the profile library and removed the requirement
to pass in a copy of the configuration when initializing the GettextMixin
object. It can now read the configuration directly. I had to add in
*args as a placeholder, though, since every plugin calls this library and
I didn't want to update all of them. Again, not exactly related to the
problem at hand, more of an issue I ran into while testing.

naomi/mic.py - added some code to create a new table in audiolog which
keeps track of your training history.

naomi/plugin.py - added the STTTrainerPlugin class. Right now it is not
really populated as I'm not sure how much, if any, duplicate code can be
moved into the base definition and out of the specific objects.

naomi/pluginstore.py - added the STTTrainerPlugin as stt_trainer type, also
modified the license check to accept the "License" line in the plugin.info
file or an actual LICENSE, LICENSE.md, or LICENSE.txt as created by
GitHub.

naomi/populate.py - just a couple of lines to deal with the changes to
commandline.

naomi/profile.py - sometimes it gets confusing when passing lists or tuples
to the profile functions when the path is only one level deep. With the
settings, you have to use tuples to hold path information since only tuples
can be used as indexes for dictionaries. Unfortunately, a tuple containing
only a single value seems to be interpreted as that value. I added a check
to the start of every function taking a path as argument to check if that
path is a string, and wrap it in a list if so.

plugins/speechhandler/check_email/check_email.py - I am currently working on
adding some additional functionality to the email system. This works toward
that goal by changing the "email: imap: setting" to "email: imap: server:" and
the "email: imap_port" to "email: imap: port:"

NaomiSTTTrainer.py - implemented a plugin system to make it easier to build
plugins for training different speech to text systems based on information
in audiolog.

plugins/stt_trainer/pocketsphinx_adapt - This folder contains the
pocketsphinx_adapt stt training program.