Naomi WebRTC Chunk Length
This pull request implements a fix for sound cards that do not honor the
chunk size when set through the alsaaudio.PCM().setperiodsize(chunk_size)
function.

This specifically affects the Intel 82801AA-ICH card that VirtualBox
creates as a passthrough device from the virtual machine to the hardware,
but I have heard of others having the same trouble with certain hardware
sound cards.

In addition to providing guidance to set your input_chunksize to something
that will result in a 10, 20, or 30 ms chunk length, now the webrtcvad
plugin will attempt to "fix" a chunk which is not the correct length to
work with webrtcvad by cutting it down to the next smaller valid sample
length. So if your sound card insists on only returning chunks of length
682 (which is what VirtualBox was doing), it sees that a 20ms chunk at
16000Hz is 640bytes (16000 samples per second *.02 seconds * 2 bytes per
sample) and cuts the chunk down to the first 640 bytes before running
webrtcvad.is_speech() on it. Hopefully this will be helpful to someone.

I also fixed a small problem where populate.py was searching for
audioengine and stt plugins separate from the main plugin search, which
was causing the plugins: section of profile.py to create a small section
with an empty label that only contained them, because the way we were
searching for category specific plugins we were losing the category name.
This fixes it so you can pass a separate, optional "category" parameter
to pluginstore.detect_plugins(). This also makes the plugin_dirs
parameter optional since we generally always search the same two
locations for plugins.

Here are the files affected:

naomi/pluginstore.py - made plugin_dirs optional when instantiating a
PluginStore object. If nothing is passed in, then it defaults to
~/Naomi/plugins and ~/.naomi/plugins (or whatever the equivalents
are under the specific setup). Added an optional "category" parameter
to detect_plugins() so you can detect only a specific category of
plugins by just passing the name in without setting the specific
folder to search.

naomi/populate.py - changed the prompt for email or text message so
it didn't appear to be asking the question twice. Fixed the tts
and audioengine plugin searches so they didn't create ghost
entries in profile.yml.

plugins/stt_trainer/pocketsphinx_adapt/packetsphinx_adapt.py -
Changed to use the new PluginStore method for searching for
speechhandler plugins.

plugins/vad/snr_vad/snr_vad.py - changed the tolerance multiplier
to default to 1, but can now be set by the user through the
snr_vad: tolerance: setting in profile.yml

plugins/vad/webrtc_vad/webrtc_vad.py - made it so that webrtc_vad
can truncate the chunk size to the next smaller allowed chunk size
if alsa is overriding the requested chunk size. Still demands that
the user sets the chunk size in profile.yml as that is still
optimal.