Language Variable Bugfix
Pull Request #196 NaomiSTTTrainer added some gettext functions to
naomi/plugin.py. To make sure that language was configured, I
added the line self.language=interface.get_language().

Unfortunately, that was incompatible with witai-stt. That plugin
defined it's own language property using @property but did not
set a setter, so it was impossible to set that property. The
solution was to either give that property a setter, delete the
getter, or just change the line in plugin.py to
language=interface.get_language()
since I'm not really using the language variable outsite of
the init anyway. WitAI also does not appear to be using its
"language" variable.

Also included in this bugfix is a fix for espeak. If you had
not set a voice for that TTS plugin, the get_voices() method
was trying to figure out the best voice for you based on your
language, but was trying to use regex on a binary object
returned by the subprocess method, which was blowing up. This
seems to be an aftereffect of the conversion to Python3.