VAD Visualization
This pull request is to add a display to the snr_vad voice
activity detection engine, which is the default for Naomi.

Sometimes it is difficult to tell what Naomi is doing as far
as listening. Sometimes when it does not respond right away it
is because it didn't realize you were speaking. Sometimes it is
because it has not realized that you have stopped speaking yet.
And sometimes it just didn't recognize any words in what it
heard. This provides some visual feedback. I suggest using it
together with the print_transcript and passive_listen options
for the best overall experience.

The goal is not to rely on this to provide visual feedback, but
to provide feedback so that we can improve Naomi's responsiveness
to the point where the feedback is no longer necessary.

naomi/commandline.py - Added a new "println" function that is used by the
snr_vad and naomi_transcript when printing to the display. Rather than
constantly adding new lines and scrolling up, this tries to print the full
display width, adding spaces after whatever is being printed to cover the
current line.

If you want to print to the screen and there might be something on that
line already, then you should use the commandline.println() function.
Because this is used to print to the line and then go back to print on
the same line again, you should put a "\n" line feed at the end of any
string you want to stay up on the screen.

naomi/mic.py - switched to using the new println function when printing
the transcript of your conversation with naomi to the screen.

naomi/plugin.py - added "recording" as a boolean parameter to the call
to voice_detected, so that I could display a character to show whether
Naomi is recording or not.

plugins/vad/snr_vad/snr_vad.py - changed timeout to 10 - this means
Naomi will record at least 10 frames after it stops detecting the user
speaking. Since each frame is approx .03 seconds, that means it will
keep about 1/3 second of audio after the voice drops below threshold.
I noticed that sometimes the audio was getting cut off at the end.
Also changed the _voice_detected method to use *args and **kwargs so
I could receive the "recording" parameter if passed in. Finally, this
is where the actual display is built and shown.

plugins/vad/webrtc_vad/webrtc_vad.py - changed default timeout to 10
here also. Also changed the _voice_detected method to use *args and
**kwargs to bring it in line with snr_vad.py, even though I'm not
adding a display here, but I should be able to at some point. In this
case I don't have so much information available about what the vad
engine is doing, but I could at least create a display of showing
when recording begins and how long it goes before ending.